Break down externalTypes/types.ts into more appropriately organized places
