+++ packages/server/src/ApolloServer.ts
@@ -1,78 +1,75 @@
+import type { Logger } from '@apollo/utils.logger';
 import { addMocksToSchema } from '@graphql-tools/mock';
 import { makeExecutableSchema } from '@graphql-tools/schema';
-import loglevel from 'loglevel';
-import Negotiator from 'negotiator';
-import {
-  GraphQLSchema,
-  GraphQLError,
-  ValidationContext,
-  FieldDefinitionNode,
-  DocumentNode,
-  ParseOptions,
-  print,
-  assertValidSchema,
-  GraphQLFormattedError,
-  GraphQLFieldResolver,
-} from 'graphql';
 import resolvable, { Resolvable } from '@josephg/resolvable';
 import {
   InMemoryLRUCache,
   KeyValueCache,
   PrefixingKeyValueCache,
 } from 'apollo-server-caching';
+import {
+  assertValidSchema,
+  DocumentNode,
+  FieldDefinitionNode,
+  GraphQLError,
+  GraphQLFieldResolver,
+  GraphQLFormattedError,
+  GraphQLSchema,
+  ParseOptions,
+  print,
+  ValidationContext,
+} from 'graphql';
+import loglevel from 'loglevel';
+import Negotiator from 'negotiator';
+import * as uuid from 'uuid';
+import { newCachePolicy } from './cachePolicy';
+import { ApolloConfig, determineApolloConfig } from './config';
+import { formatApolloErrors } from './errors';
 import type {
   ApolloServerPlugin,
-  GraphQLServiceContext,
-  GraphQLServerListener,
-  LandingPage,
   BaseContext,
-  GraphQLResponse,
+  ContextFunction,
   GraphQLExecutor,
+  GraphQLResponse,
+  GraphQLServerListener,
+  GraphQLServiceContext,
   HTTPGraphQLRequest,
   HTTPGraphQLResponse,
-  WithRequired,
-  ContextFunction,
-} from './externalTypes';
-import type { Logger } from '@apollo/utils.logger';
-
-import type {
-  ApolloServerOptions,
-  DocumentStore,
-  PersistedQueryOptions,
+  LandingPage,
   PluginDefinition,
-  ApolloServerOptionsWithStaticSchema,
-} from './types';
-
+} from './externalTypes';
+import { runPotentiallyBatchedHttpQuery } from './httpBatching';
+import { InternalPluginId, pluginIsInternal } from './internalPlugin';
 import {
-  processGraphQLRequest,
-  GraphQLRequestContext,
-  GraphQLRequest,
-  APQ_CACHE_PREFIX,
-} from './requestPipeline';
-
-import isNodeLike from './utils/isNodeLike';
-import { ApolloConfig, determineApolloConfig } from './config';
-import {
-  ApolloServerPluginSchemaReporting,
-  ApolloServerPluginSchemaReportingOptions,
-  ApolloServerPluginInlineTrace,
-  ApolloServerPluginUsageReporting,
   ApolloServerPluginCacheControl,
+  ApolloServerPluginInlineTrace,
   ApolloServerPluginLandingPageLocalDefault,
   ApolloServerPluginLandingPageProductionDefault,
+  ApolloServerPluginSchemaReporting,
+  ApolloServerPluginSchemaReportingOptions,
+  ApolloServerPluginUsageReporting,
 } from './plugin';
-import { InternalPluginId, pluginIsInternal } from './internalPlugin';
-import { newCachePolicy } from './cachePolicy';
-import { GatewayIsTooOldError, SchemaManager } from './utils/schemaManager';
-import * as uuid from 'uuid';
+import {
+  APQ_CACHE_PREFIX,
+  GraphQLRequest,
+  GraphQLRequestContext,
+  processGraphQLRequest,
+} from './requestPipeline';
 import {
   cloneObject,
   HeaderMap,
   HttpQueryError,
   prettyJSONStringify,
 } from './runHttpQuery';
-import { runPotentiallyBatchedHttpQuery } from './httpBatching';
-import { formatApolloErrors } from './errors';
+import type {
+  ApolloServerOptions,
+  ApolloServerOptionsWithStaticSchema,
+  DocumentStore,
+  PersistedQueryOptions,
+  WithRequired,
+} from './types';
+import isNodeLike from './utils/isNodeLike';
+import { GatewayIsTooOldError, SchemaManager } from './utils/schemaManager';
 
 const NoIntrospection = (context: ValidationContext) => ({
   Field(node: FieldDefinitionNode) {
