+++ packages/server/src/externalTypes/index.ts
@@ -1,9 +1,40 @@
-// FIXME: this goes away once types all find their proper homes
-// These are the types that used to live in `apollo-server-plugin-base`.
-export * from './plugins';
-
-// These are the types that used to live in `@apollo/server-types`.
-export * from './types';
-
+export {
+  CacheAnnotation,
+  CacheHint,
+  CachePolicy,
+  CacheScope,
+} from './cacheControl';
 export { BaseContext, ContextFunction } from './context';
+export {
+  GraphQLRequest,
+  GraphQLRequestContext,
+  GraphQLResponse,
+} from './graphql';
 export { HTTPGraphQLRequest, HTTPGraphQLResponse } from './http';
+export {
+  ApolloServerPlugin,
+  GraphQLFieldResolverParams,
+  GraphQLRequestExecutionListener,
+  GraphQLRequestListener,
+  GraphQLRequestListenerDidResolveField,
+  GraphQLRequestListenerExecutionDidEnd,
+  GraphQLRequestListenerParsingDidEnd,
+  GraphQLRequestListenerValidationDidEnd,
+  GraphQLSchemaContext,
+  GraphQLServerListener,
+  GraphQLServiceContext,
+  LandingPage,
+  PluginDefinition,
+} from './plugins';
+export {
+  GraphQLExecutionResult,
+  GraphQLExecutor,
+  GraphQLRequestContextDidEncounterErrors,
+  GraphQLRequestContextDidResolveOperation,
+  GraphQLRequestContextDidResolveSource,
+  GraphQLRequestContextExecutionDidStart,
+  GraphQLRequestContextParsingDidStart,
+  GraphQLRequestContextResponseForOperation,
+  GraphQLRequestContextValidationDidStart,
+  GraphQLRequestContextWillSendResponse,
+} from './requestPipeline';
